Use legacy behavior if Python extension hasn't been upgraded
